Using long instead of int for scores is safer

+++ walls-of-doom/record.h
@@ -14,7 +14,7 @@
  */
 typedef struct Record {
   char name[MAXIMUM_PLAYER_NAME_SIZE];
-  int score;
+  long score;
 } Record;
 
 /**
